Added saving graph to file, and added main method.
So the generate_graph function will now take a post_id as input, which it'll use to assign a unique filename to the graph, which can also be stored to the database.

Also added a main method so the file doesn't run code to test the functionality when importing.

Also went with horizontal gridlines.